[F] Use journal schema description for about content Resolves WDP-483
