+++ packages/frontend/components/atomic/FullTextProperty/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FullTextProperty";
