+++ packages/frontend/components/composed/journal/JournalContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./JournalContent";
