--- packages/frontend/components/composed/journal/JournalInfo/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./JournalInfo";
