+++ packages/frontend/locales/en.json
@@ -99,7 +99,8 @@
       "online_version_label": "Online Version",
       "contributions_header": "Contributions",
       "back_to_contributors": "Back to Contributors",
-      "current_issue": "Current $t(schema.nglp.journal_issue)"
+      "current_issue": "Current $t(schema.nglp.journal_issue)",
+      "announcements_header": "Announcements"
     },
     "nav": {
       "home": "Home",
